requirements.txt
# FoodVisionAI - Requirements
# Python 3.10+ recommended

# Core ML/DL
tensorflow>=2.15.0,<2.18.0
tensorflow-datasets>=4.9.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data Science
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Model Evaluation
scikit-learn>=1.3.0

# UI (for later use)
streamlit>=1.28.0
gradio>=4.0.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0